+++ folio-spring-system-user/src/main/java/org/folio/spring/service/PrepareSystemUserService.java
@@ -2,15 +2,19 @@ package org.folio.spring.service;
 
 import static org.springframework.util.CollectionUtils.isEmpty;
 
+import feign.FeignException;
+import java.io.IOException;
+import java.io.UncheckedIOException;
 import java.nio.charset.StandardCharsets;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.io.IOUtils;
+import org.folio.spring.FolioExecutionContext;
 import org.folio.spring.client.AuthnClient;
 import org.folio.spring.client.AuthnClient.UserCredentials;
 import org.folio.spring.client.PermissionsClient;
@@ -29,7 +33,9 @@ public class PrepareSystemUserService {
 
   public static final String SYSTEM_USER_TYPE = "system";
 
+  private final FolioExecutionContext folioExecutionContext;
   private final SystemUserProperties systemUserProperties;
+  private final SystemUserService systemUserService;
 
   private UsersClient usersClient;
   private AuthnClient authnClient;
@@ -41,23 +47,61 @@ public class PrepareSystemUserService {
       return;
     }
 
-    log.info("Preparing system user...");
-    var folioUser = getFolioUser(systemUserProperties.username());
-    var userId = folioUser.map(User::id)
-        .orElse(UUID.randomUUID().toString());
+    try {
+      log.info("Preparing system user...");
+
+      String userId = getOrCreateSystemUser();
+      assignPermissions(userId);
+      saveCredentials(userId);
+
+      // a nice sanity check, to fail sooner and ensure later logins will go smoothly
+      systemUserService.authSystemUser(folioExecutionContext.getTenantId(),
+                                       systemUserProperties.username(), systemUserProperties.password());
+      log.info("System user authenticated successfully");
+    } catch (RuntimeException e) {
+      log.error("Unexpected error while preparing system user with username={}:", systemUserProperties.username(), e);
+      throw e;
+    }
+
+    log.info("Preparing system user is completed!");
+  }
+
+  /**
+   * Gets or creates the system user (reactivating if needed).
+   *
+   * @return the system user's ID
+   */
+  private String getOrCreateSystemUser() {
+    Optional<User> folioUser = systemUserService.getFolioUser(systemUserProperties.username());
+    String userId = folioUser.map(User::getId).orElse(UUID.randomUUID().toString());
 
     if (folioUser.isPresent()) {
-      log.info("System user already exists");
-      addPermissions(userId);
-      deleteCredentials(userId);
-      saveCredentials();
+      User user = folioUser.get();
+      log.info("Found existing system user, id={}", user.getId());
+
+      if (!user.isActive()) {
+        log.info("System user is inactive, attempting to mark active...");
+        user = user.toBuilder().active(true).expirationDate(null).build();
+        try {
+          usersClient.updateUser(user);
+        } catch (FeignException e) {
+          log.error("System user was unable to be marked active, so we are failing module upgrade.");
+          log.error("Attempting to mark system user active resulted in the following error:", e);
+          throw e;
+        }
+      }
     } else {
-      log.info("No system user exist, creating...");
+      log.info("Could not find a system user with username={}, creating a new one with id={}...",
+                systemUserProperties.username(), userId);
       createFolioUser(userId);
-      assignPermissions(userId);
-      saveCredentials();
     }
-    log.info("System user has been created");
+
+    return userId;
+  }
+
+  private void createFolioUser(String id) {
+    final User user = prepareUserObject(id);
+    usersClient.createUser(user);
   }
 
   public void deleteCredentials(String userId) {
@@ -66,59 +110,75 @@ public class PrepareSystemUserService {
     log.info("Removed credentials for user {}.", userId);
   }
 
-  public Optional<User> getFolioUser(String username) {
-    var users = usersClient.query("username==" + username);
-    return (users == null || users.getResult() == null) ? Optional.empty() : users.getResult().stream().findFirst();
+  /** Save the credentials for a user, removing an existing login if necessary. */
+  private void saveCredentials(String id) {
+    saveCredentials(id, true);
   }
 
-  private void createFolioUser(String id) {
-    final var user = prepareUserObject(id);
-    usersClient.saveUser(user);
-  }
+  private void saveCredentials(String id, boolean clearExisting) {
+    try {
+      authnClient.saveCredentials(new UserCredentials(systemUserProperties.username(),
+                                                      systemUserProperties.password()));
 
-  private void saveCredentials() {
-    authnClient.saveCredentials(new UserCredentials(systemUserProperties.username(), systemUserProperties.password()));
+      log.info("Saved credentials for user with username={}", systemUserProperties.username());
+    } catch (FeignException.UnprocessableEntity e) {
+      if (!clearExisting) {
+        throw e;
+      }
 
-    log.info("Saved credentials for user: [{}]", systemUserProperties.username());
+      log.warn("Credentials already exist for user with username={}, removing them and re-adding...",
+               systemUserProperties.username());
+
+      deleteCredentials(id);
+      saveCredentials(id, false);
+    }
   }
 
   private void assignPermissions(String userId) {
     List<String> permissionsToAssign = getResourceLines(systemUserProperties.permissionsFilePath());
+    log.info("Found {} permissions to assign: {}", permissionsToAssign.size(), permissionsToAssign);
 
     if (isEmpty(permissionsToAssign)) {
-      throw new IllegalStateException("No permissions found to assign to user with id: " + userId);
+      throw log.throwing(
+        new IllegalStateException("No permissions found to assign to system user; maybe the permissions file is empty?")
+      );
     }
 
-    var permissions = new Permissions(UUID.randomUUID().toString(), userId, permissionsToAssign);
+    try {
+      List<String> existingPermissions = permissionsClient.getUserPermissions(userId).getResult();
+      log.info("System user currently has {} permissions", existingPermissions.size());
 
-    permissionsClient.assignPermissionsToUser(permissions);
-    log.info("Permissions assigned to system user: [{}]", permissionsToAssign);
-  }
+      Set<String> permissionsToAdd = new HashSet<>(permissionsToAssign);
+      existingPermissions.forEach(permissionsToAdd::remove);
 
-  private void addPermissions(String userId) {
-    var permissionsToAssign = getResourceLines(systemUserProperties.permissionsFilePath());
+      log.info("Adding {} permissions to system user: {}", permissionsToAdd.size(), permissionsToAdd);
 
-    if (isEmpty(permissionsToAssign)) {
-      throw new IllegalStateException("No permissions found to assign to user with id: " + userId);
+      permissionsToAdd.forEach(permission -> permissionsClient.addPermission(userId, new Permission(permission)));
+    } catch (FeignException.NotFound | NullPointerException e) {
+      UUID newPermissionsId = UUID.randomUUID();
+      log.warn("No permissions record found for system user, creating a fresh one with id={}...", newPermissionsId);
+      Permissions permissions = new Permissions(UUID.randomUUID().toString(), userId, permissionsToAssign);
+
+      permissionsClient.assignPermissionsToUser(permissions);
     }
 
-    var permissionsToAdd = new HashSet<>(permissionsToAssign);
-    permissionsClient.getUserPermissions(userId).getResult().forEach(permissionsToAdd::remove);
-
-    permissionsToAdd.forEach(permission ->
-        permissionsClient.addPermission(userId, new Permission(permission)));
-    log.info("Permissions assigned to system user: [{}]", permissionsToAdd);
+    log.info("System user permission assignment successful");
   }
 
   private User prepareUserObject(String id) {
-    return new User(id, systemUserProperties.username(), SYSTEM_USER_TYPE,  true,
-        new User.Personal(systemUserProperties.lastname()));
+    return User.builder().id(id).username(systemUserProperties.username()).type(SYSTEM_USER_TYPE).active(true)
+              .personal(new User.Personal(systemUserProperties.lastname())).build();
   }
 
-  @SneakyThrows
   private List<String> getResourceLines(String permissionsFilePath) {
-    var resource = new ClassPathResource(permissionsFilePath);
-    return IOUtils.readLines(resource.getInputStream(), StandardCharsets.UTF_8);
+    try {
+      ClassPathResource resource = new ClassPathResource(permissionsFilePath);
+      List<String> lines = IOUtils.readLines(resource.getInputStream(), StandardCharsets.UTF_8);
+      lines.removeIf(String::isBlank);
+      return lines;
+    } catch (IOException | UncheckedIOException e) {
+      throw log.throwing(new IllegalArgumentException("Unable to open permissions file " + permissionsFilePath, e));
+    }
   }
 
   @Autowired(required = false)