+++ folio-spring-system-user/src/test/java/org/folio/spring/context/ExecutionContextBuilderTest.java
@@ -38,12 +38,13 @@ class ExecutionContextBuilderTest {
         .okapiUrl("okapi").tenantId("tenant")
         .userId(userId.toString())
         .build();
-    var context = builder.forSystemUser(systemUser);
+    var context = builder.forSystemUser(systemUser, null);
 
     assertThat(context.getTenantId()).isEqualTo("tenant");
     assertThat(context.getToken()).isEqualTo("token");
     assertThat(context.getOkapiUrl()).isEqualTo("okapi");
     assertThat(context.getUserId()).isEqualTo(userId);
+    assertThat(context.getRequestId()).isNullOrEmpty();
 
     assertThat(context.getAllHeaders()).isNotNull();
     assertThat(context.getOkapiHeaders()).isNotNull().hasSize(4);
@@ -58,11 +59,12 @@ class ExecutionContextBuilderTest {
         .userId(null)
         .build();
 
-    var context = builder.forSystemUser(systemUser);
+    var context = builder.forSystemUser(systemUser, null);
 
     assertThat(context.getTenantId()).isEqualTo(EMPTY);
     assertThat(context.getToken()).isEqualTo(EMPTY);
     assertThat(context.getOkapiUrl()).isEqualTo(EMPTY);
+    assertThat(context.getRequestId()).isNullOrEmpty();
 
     assertThat(context.getAllHeaders()).isNotNull();
     assertThat(context.getOkapiHeaders()).isNotNull().isEmpty();
@@ -80,6 +82,7 @@ class ExecutionContextBuilderTest {
     assertThat(context.getTenantId()).isEqualTo(tenantId);
     assertThat(context.getToken()).isEqualTo(EMPTY);
     assertThat(context.getOkapiUrl()).isEqualTo(okapiUrl);
+    assertThat(context.getRequestId()).isNullOrEmpty();
 
     assertThat(context.getAllHeaders()).hasSize(2);
     assertThat(context.getOkapiHeaders()).hasSize(2);